Avoid any potential out-of-bounds indexing in neasyf_type
Allocate a local variable with the same type as the input, but with
exactly one element. This way we can be sure we can index it correctly
to call the scalar version.

This shouldn't be a problem now that we explicitly disallow zero-sized
dimensions, but it's nice to be safe in case the user created the
dimensions some other way.